elements in enum may be of any type
