+++ lib/v3.0/schema/parse.js
@@ -9,31 +9,9 @@ const {
 const {
   parseReferenceObject
 } = require('../parseReferenceObject')
-
-// https://datatracker.ietf.org/doc/html/draft-wright-json-schema-00#section-4.2
-function valueToLiteral (value) {
-  if (value === null) {
-    return ts.factory.createNull()
-  }
-
-  if (value === true) {
-    return ts.factory.createTrue()
-  }
-
-  if (value === false) {
-    return ts.factory.createFalse()
-  }
-
-  if (typeof value === 'number') {
-    return ts.factory.createNumericLiteral(value)
-  }
-
-  if (typeof value === 'string') {
-    return ts.factory.createStringLiteral(value)
-  }
-
-  throw new Error(`unexpected value type. [${typeof value}]`)
-}
+const {
+  valueToLiteral
+} = require('../valueToLiteral')
 
 // https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#schema-object
 function parseSchemaObjectProperties (
