+++ lib/v3.0/type/parse.js
@@ -10,6 +10,9 @@ const {
 const {
   parseReferenceObject
 } = require('../parseReferenceObject')
+const {
+  valueToLiteral
+} = require('../valueToLiteral')
 
 // https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#data-types
 function toOASTypeNode (type) {
@@ -58,7 +61,7 @@ function toTypeNode (
   if (Array.isArray(obj[FIELD_NAME.ENUM])) {
     return ts.factory.createUnionTypeNode(
       obj[FIELD_NAME.ENUM].map(e => ts.factory.createLiteralTypeNode(
-        ts.factory.createStringLiteral(e)
+        valueToLiteral(e)
       ))
     )
   }
