+++ lib/v3.1/schema/parse.js
@@ -10,31 +10,9 @@ const {
 const {
   parseReferenceObject
 } = require('../parseReferenceObject')
-
-// https://datatracker.ietf.org/doc/html/draft-bhutton-json-schema-00#section-4.2.1
-function valueToLiteral (value) {
-  if (value === null) {
-    return ts.factory.createNull()
-  }
-
-  if (value === true) {
-    return ts.factory.createTrue()
-  }
-
-  if (value === false) {
-    return ts.factory.createFalse()
-  }
-
-  if (typeof value === 'number') {
-    return ts.factory.createNumericLiteral(value)
-  }
-
-  if (typeof value === 'string') {
-    return ts.factory.createStringLiteral(value)
-  }
-
-  throw new Error(`unexpected value type. [${typeof value}]`)
-}
+const {
+  valueToLiteral
+} = require('../valueToLiteral')
 
 // https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#schema-object
 function parseSchemaObjectProperties (
