+++ lib/v3.1/type/parse.js
@@ -11,6 +11,9 @@ const {
 const {
   parseReferenceObject
 } = require('../parseReferenceObject')
+const {
+  valueToLiteral
+} = require('../valueToLiteral')
 
 // https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#data-types
 function toOASTypeNode (type) {
@@ -67,7 +70,7 @@ function toTypeNode (
   if (Array.isArray(obj[FIELD_NAME.ENUM])) {
     return ts.factory.createUnionTypeNode(
       obj[FIELD_NAME.ENUM].map(e => ts.factory.createLiteralTypeNode(
-        ts.factory.createStringLiteral(e)
+        valueToLiteral(e)
       ))
     )
   }
